requirements.txt
Flask
tensorflow
numpy
Pillow
google-cloud-storage
werkzeug
